+++ src/ckanapi_harvesters/harvesters/file_formats/auxiliary_defs.py
@@ -5,7 +5,9 @@ Auxiliary definitions
 """
 from enum import IntEnum
 
+
 class GeoFileStoreEpsgConversion(IntEnum):
     CsvWkb = 0
     ShapefileProjection = 2
     ShapefileAsIs = 3
+